fixing data type waiting duration

+++ features/content/play/src/main/java/com/tokopedia/play/view/uimodel/PlayUpcomingUiModel.kt
@@ -10,11 +10,11 @@ data class PlayUpcomingUiModel(
     val coverUrl: String = "",
     val startTime: String = "",
     val isAlreadyLive: Boolean = false,
-    val refreshWaitingDuration: Long = REFRESH_WAITING_DURATION,
+    val refreshWaitingDuration: Int = REFRESH_WAITING_DURATION,
 ) {
     companion object {
         const val COMING_SOON = "COMING_SOON"
 
-        const val REFRESH_WAITING_DURATION = 5000L
+        const val REFRESH_WAITING_DURATION = 5000
     }
 }
