+++ features/content/play/src/main/java/com/tokopedia/play/view/uimodel/recom/PlayStatusInfoUiModel.kt
@@ -10,7 +10,7 @@ data class PlayStatusInfoUiModel(
         val bannedModel: PlayBannedUiModel,
         val freezeModel: PlayFreezeUiModel,
         val shouldAutoSwipeOnFreeze: Boolean,
-        val waitingDuration: Long,
+        val waitingDuration: Int,
 )
 
 data class PlayBannedUiModel(
