+++ features/content/play/src/main/java/com/tokopedia/play/view/uimodel/state/PlayUpcomingUiState.kt
@@ -18,7 +18,7 @@ data class PlayUpcomingGeneralInfo(
     val title: String,
     val coverUrl: String,
     val startTime: String,
-    val waitingDuration: Long,
+    val waitingDuration: Int,
 )
 
 sealed class PlayUpcomingState {
